goodhabit/requirements.txt
amqp==5.2.0
anyio==4.4.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.1
cron-descriptor==1.4.5
Django==5.1.1
django-celery-beat==2.7.0
django-cors-headers==4.0.0
django-timezone-field==7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
inflection==0.5.1
iniconfig==2.0.0
kombu==5.4.1
packaging==24.1
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytest==8.3.3
pytest-cov==5.0.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
python-telegram-bot==21.5
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
